requirements.txt
fastapi~=0.109.1
uvicorn~=0.15.0
tortoise-orm~=0.17.8
tortoise~=0.1.1
asyncpg~=0.24.0
requests~=2.32.3
pycryptodome~=3.19.0
minio~=7.2.15
python-jose~=3.4.0
PyJWT~=2.3.0
pydantic~=2.10.6